+++ src/game/units/wizard/archetypes/shepherd/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod constants;
+mod plugin;
+
+pub use plugin::ShepherdPlugin;